+++ core/responseFnURL.go
@@ -7,36 +7,33 @@ import (
 	"encoding/base64"
 	"errors"
 	"net/http"
+	"strings"
 	"unicode/utf8"
 
 	"github.com/aws/aws-lambda-go/events"
 )
 
-// ProxyResponseWriterFunctionURL implements http.ResponseWriter and adds the method
+// FunctionUrlResponseWriter implements http.ResponseWriter and adds the method
 // necessary to return an events.LambdaFunctionURLResponse object
-type ProxyResponseWriterFunctionURL struct {
-	status    int
+type FunctionUrlResponseWriter struct {
 	headers   http.Header
 	body      bytes.Buffer
+	status    int
 	observers []chan<- bool
 }
 
-// Ensure implementation satisfies http.ResponseWriter interface
-var (
-	_ http.ResponseWriter = &ProxyResponseWriterFunctionURL{}
-)
-
-// NewProxyResponseWriterFnURL returns a new ProxyResponseWriterFunctionURL object.
-// The object is initialized with an empty map of headers and a status code of -1
-func NewProxyResponseWriterFnURL() *ProxyResponseWriterFunctionURL {
-	return &ProxyResponseWriterFunctionURL{
+// NewFunctionUrlResponseWriter returns a new FunctionUrlResponseWriter object.
+// The object is initialized with an empty map of headers and a
+// status code of -1
+func NewFunctionUrlResponseWriter() *FunctionUrlResponseWriter {
+	return &FunctionUrlResponseWriter{
 		headers:   make(http.Header),
 		status:    defaultStatusCode,
 		observers: make([]chan<- bool, 0),
 	}
 }
 
-func (r *ProxyResponseWriterFunctionURL) CloseNotify() <-chan bool {
+func (r *FunctionUrlResponseWriter) CloseNotify() <-chan bool {
 	ch := make(chan bool, 1)
 
 	r.observers = append(r.observers, ch)
@@ -44,21 +41,21 @@ func (r *ProxyResponseWriterFunctionURL) CloseNotify() <-chan bool {
 	return ch
 }
 
-func (r *ProxyResponseWriterFunctionURL) notifyClosed() {
+func (r *FunctionUrlResponseWriter) notifyClosed() {
 	for _, v := range r.observers {
 		v <- true
 	}
 }
 
 // Header implementation from the http.ResponseWriter interface.
-func (r *ProxyResponseWriterFunctionURL) Header() http.Header {
+func (r *FunctionUrlResponseWriter) Header() http.Header {
 	return r.headers
 }
 
 // Write sets the response body in the object. If no status code
 // was set before with the WriteHeader method it sets the status
 // for the response to 200 OK.
-func (r *ProxyResponseWriterFunctionURL) Write(body []byte) (int, error) {
+func (r *FunctionUrlResponseWriter) Write(body []byte) (int, error) {
 	if r.status == defaultStatusCode {
 		r.status = http.StatusOK
 	}
@@ -76,19 +73,19 @@ func (r *ProxyResponseWriterFunctionURL) Write(body []byte) (int, error) {
 
 // WriteHeader sets a status code for the response. This method is used
 // for error responses.
-func (r *ProxyResponseWriterFunctionURL) WriteHeader(status int) {
+func (r *FunctionUrlResponseWriter) WriteHeader(status int) {
 	r.status = status
 }
 
 // GetProxyResponse converts the data passed to the response writer into
-// an events.LambdaFunctionURLResponse object.
+// an events.APIGatewayProxyResponse object.
 // Returns a populated proxy response object. If the response is invalid, for example
 // has no headers or an invalid status code returns an error.
-func (r *ProxyResponseWriterFunctionURL) GetProxyResponse() (events.LambdaFunctionURLResponse, error) {
+func (r *FunctionUrlResponseWriter) GetProxyResponse() (events.LambdaFunctionURLResponse, error) {
 	r.notifyClosed()
 
 	if r.status == defaultStatusCode {
-		return events.LambdaFunctionURLResponse{}, errors.New("status code not set on response")
+		return events.LambdaFunctionURLResponse{}, errors.New("Status code not set on response")
 	}
 
 	var output string
@@ -104,8 +101,14 @@ func (r *ProxyResponseWriterFunctionURL) GetProxyResponse() (events.LambdaFuncti
 	}
 
 	headers := make(map[string]string)
-	for h, v := range r.Header() {
-		headers[h] = v[0]
+	cookies := make([]string, 0)
+
+	for headerKey, headerValue := range http.Header(r.headers) {
+		if strings.EqualFold("set-cookie", headerKey) {
+			cookies = append(cookies, headerValue...)
+			continue
+		}
+		headers[headerKey] = strings.Join(headerValue, ",")
 	}
 
 	return events.LambdaFunctionURLResponse{
@@ -113,5 +116,6 @@ func (r *ProxyResponseWriterFunctionURL) GetProxyResponse() (events.LambdaFuncti
 		Headers:         headers,
 		Body:            output,
 		IsBase64Encoded: isBase64,
+		Cookies:         cookies,
 	}, nil
 }
