+++ core/typesFnURL.go
@@ -6,7 +6,6 @@ import (
 	"github.com/aws/aws-lambda-go/events"
 )
 
-// GatewayTimeoutFnURL returns a dafault Gateway Timeout (504) response
-func GatewayTimeoutFnURL() events.LambdaFunctionURLResponse {
+func FunctionURLTimeout() events.LambdaFunctionURLResponse {
 	return events.LambdaFunctionURLResponse{StatusCode: http.StatusGatewayTimeout}
 }
